# Conflicts:
#	src/YesSql.Core/Services/DefaultQuery.cs
#	src/YesSql.Provider.PostgreSql/PostgreSqlDialect.cs
#	src/YesSql.Provider.SqlServer/SqlServerDialect.cs
#	test/YesSql.Tests/SqlServer2019Tests.cs

+++ src/YesSql.Core/SessionState.cs
@@ -25,8 +25,8 @@ namespace YesSql
         public HashSet<object> Tracked => _tracked ??= new HashSet<object>();
 
         // ids of entities that are checked for concurrency
-        internal HashSet<int> _concurrent;
-        public HashSet<int> Concurrent => _concurrent ??= new HashSet<int>();
+        internal HashSet<long> _concurrent;
+        public HashSet<long> Concurrent => _concurrent ??= new HashSet<long>();
 
         // entities that need to be deleted in the next flush
         internal HashSet<object> _deleted;
